+++ src/divider/__test__/index.test.js
@@ -13,6 +13,19 @@ describe('Divider', () => {
     rootPath: path.resolve(__dirname, '../..'),
   });
 
+  it(`:base`, () => {
+    const id = simulate.load({
+      template: `<t-divider class="divider"></t-divider>`,
+      usingComponents: {
+        't-divider': divider,
+      },
+    });
+    const comp = simulate.render(id);
+    comp.attach(document.createElement('parent-wrapper'));
+
+    expect(comp.toJSON()).toMatchSnapshot();
+  });
+
   textAlign.forEach((align) => {
     it(`:align ${align} render correctly`, () => {
       const id = simulate.load({