+++ src/icon/icon.ts
@@ -1,44 +1,53 @@
 import { SuperComponent, wxComponent } from '../common/src/index';
 import config from '../common/config';
 import props from './props';
+import { styles, addUnit } from '../common/utils';
 
 const { prefix } = config;
 const name = `${prefix}-icon`;
-
-const sizeKeywordMap = {
-  xs: '24rpx',
-  small: '28rpx',
-  middle: '32rpx',
-  large: '36rpx',
-  xl: '40rpx',
-};
 
 @wxComponent()
 export default class Icon extends SuperComponent {
   behaviors: ['wx://form-field-icon'];
 
-  externalClasses = ['t-class'];
+  externalClasses = [`${prefix}-class`];
 
   properties = props;
 
   data = {
+    componentPrefix: prefix,
     classPrefix: name,
-    fontSize: '',
+    isImage: false,
+    iconStyle: undefined,
   };
 
   observers = {
-    size(val) {
-      let fontSize = val;
-      if (Object.prototype.hasOwnProperty.call(sizeKeywordMap, val)) {
-        fontSize = sizeKeywordMap[val];
-      }
-      this.setData({ fontSize });
+    'name, color, size, customStyle'() {
+      this.setIconStyle();
     },
   };
 
   methods = {
     onTap(event: any) {
       this.triggerEvent('click', event.detail);
+    },
+
+    setIconStyle() {
+      const { name, color, size, customStyle } = this.properties;
+      const isImage = name.indexOf('/') !== -1;
+      const sizeValue = addUnit(size);
+      const sizeStyle = isImage ? { width: sizeValue, height: sizeValue } : {};
+      const colorStyle = color ? { color: color } : {};
+      const fontStyle = size ? { 'font-size': sizeValue } : {};
+      this.setData({
+        isImage,
+        iconStyle:
+          styles({
+            ...colorStyle,
+            ...fontStyle,
+            ...sizeStyle,
+          }) + customStyle,
+      });
     },
   };
 }