+++ src/image/image.ts
@@ -37,17 +37,17 @@ export default class Image extends SuperComponent {
     },
   };
 
-  onLoaded(e: any) {
-    const sdkVersion = wx.getSystemInfoSync().SDKVersion;
-    const versionArray = sdkVersion.split('.').map((v) => parseInt(v, 10));
-    // 版本号低于2.10.3时组件内部实现heightFix模式
-    if (
-      versionArray[0] < 2 ||
-      (versionArray[0] === 2 && versionArray[1] < 10) ||
-      (versionArray[0] === 2 && versionArray[1] === 10 && versionArray[2] < 3)
-    ) {
-      const mode = this.properties.mode as any as string;
-      if (mode === 'heightFix') {
+  methods = {
+    onLoaded(e: any) {
+      const sdkVersion = wx.getSystemInfoSync().SDKVersion;
+      const versionArray = sdkVersion.split('.').map((v) => parseInt(v, 10));
+      const { mode } = this.properties;
+      const isInCompatible =
+        versionArray[0] < 2 ||
+        (versionArray[0] === 2 && versionArray[1] < 10) ||
+        (versionArray[0] === 2 && versionArray[1] === 10 && versionArray[2] < 3);
+      // 版本号低于2.10.3时组件内部实现heightFix模式
+      if (mode === 'heightFix' && isInCompatible) {
         // 实现heightFix模式，保持高度和宽高比，设置对应的宽度
         const { height: picHeight, width: picWidth } = e.detail;
         const query = this.createSelectorQuery();
@@ -60,33 +60,33 @@ export default class Image extends SuperComponent {
           })
           .exec();
       }
-    }
-    this.setData({
-      isLoading: false,
-      isFailed: false,
-    });
-    this.triggerEvent('load', e.detail);
-  }
-
-  onLoadError(e: any) {
-    this.setData({
-      isLoading: false,
-      isFailed: true,
-    });
-    this.triggerEvent('error', e.detail);
-  }
-
-  update() {
-    const { src } = this.properties as any;
-    this.preSrc = src;
-    if (!src) {
-      // 链接为空时直接触发加载失败
-      this.onLoadError({ errMsg: '图片链接为空' });
-    } else {
       this.setData({
-        isLoading: true,
+        isLoading: false,
         isFailed: false,
       });
-    }
-  }
+      this.triggerEvent('load', e.detail);
+    },
+
+    onLoadError(e: any) {
+      this.setData({
+        isLoading: false,
+        isFailed: true,
+      });
+      this.triggerEvent('error', e.detail);
+    },
+
+    update() {
+      const { src } = this.properties as any;
+      this.preSrc = src;
+      if (!src) {
+        // 链接为空时直接触发加载失败
+        this.onLoadError({ errMsg: '图片链接为空' });
+      } else {
+        this.setData({
+          isLoading: true,
+          isFailed: false,
+        });
+      }
+    },
+  };
 }