+++ src/tabs/tabs.ts
@@ -28,12 +28,9 @@ export default class Tabs extends SuperComponent {
         target.index = this.children.length - 1;
         this.updateTabs();
       },
-      unlinked() {
-        this.children = this.children.map((child: any, index: number) => {
-          child.index = index;
-          return child;
-        });
-        this.updateTabs();
+      unlinked(target: WechatMiniprogram.Component.TrivialInstance) {
+        this.children = this.children.filter((item) => item.index !== target.index);
+        this.updateTabs(() => this.setTrack());
       },
     },
   };
@@ -56,6 +53,10 @@ export default class Tabs extends SuperComponent {
 
     animation(v) {
       this.setData({ animate: v });
+    },
+
+    placement() {
+      this.adjustPlacement();
     },
   };
 
@@ -81,30 +82,39 @@ export default class Tabs extends SuperComponent {
       this.setTrack();
     });
 
-    // 根据placement判断scroll-view滚动方向
-    const { placement } = this.properties;
-    let isScrollX = false;
-    let isScrollY = false;
-    if ((placement as any) === Position.top || (placement as any) === Position.bottom) {
-      isScrollX = true;
-    } else {
-      isScrollY = true;
-    }
-    this.setData({
-      isScrollX,
-      isScrollY,
-      direction: isScrollX ? 'X' : 'Y',
-    });
+    this.adjustPlacement();
     this.gettingBoundingClientRect(`.${name}`, true).then((res: any) => {
       this.containerWidth = res[0].width;
     });
   }
 
-  updateTabs() {
+  methods = {
+    adjustPlacement() {
+      // 根据placement判断scroll-view滚动方向
+      const { placement } = this.properties;
+      let isScrollX = false;
+      let isScrollY = false;
+      if ((placement as any) === Position.top || (placement as any) === Position.bottom) {
+        isScrollX = true;
+      } else {
+        isScrollY = true;
+      }
+      this.setData({
+        isScrollX,
+        isScrollY,
+        direction: isScrollX ? 'X' : 'Y',
+      });
+    },
+  };
+
+  updateTabs(cb) {
     const { children } = this;
-    this.setData({
-      tabs: children.map((child: any) => child.data),
-    });
+    this.setData(
+      {
+        tabs: children.map((child: any) => child.data),
+      },
+      cb,
+    );
     this.setCurrentIndexByName(this.properties.value);
   }
 
@@ -142,16 +152,12 @@ export default class Tabs extends SuperComponent {
 
   calcScrollOffset(
     containerWidth: number,
-    totalWidth: number,
     targetLeft: number,
     targetWidth: number,
     offset: number,
+    currentIndex: number,
   ) {
-    if (offset + targetLeft > containerWidth / 2) {
-      const maxOffset = totalWidth - containerWidth;
-      return Math.min(Math.abs(containerWidth / 2 - targetLeft - offset - targetWidth / 2), maxOffset);
-    }
-    return 0;
+    return currentIndex * targetWidth - (1 / 2) * containerWidth + targetWidth / 2;
   }
 
   setTrack() {
@@ -177,10 +183,10 @@ export default class Tabs extends SuperComponent {
         if (this.containerWidth) {
           const offset = this.calcScrollOffset(
             this.containerWidth,
-            rect.width * res.length,
             rect.left,
             rect.width,
             this.data.offset,
+            currentIndex,
           );
           this.setData({
             offset,
@@ -240,10 +246,11 @@ export default class Tabs extends SuperComponent {
 
   changeIndex(index) {
     const currentTab = this.data.tabs[index];
+    const { value, label } = currentTab;
     if (!currentTab?.disabled && index !== this.data.currentIndex) {
-      this._trigger('change', { value: currentTab.value });
+      this._trigger('change', { value, label });
     }
-    this._trigger('click', { value: currentTab.value });
+    this._trigger('click', { value, label });
   }
 
   getAvailableTabIndex(deltaX: number) {